refactor: Update comparison operators in PostgreSQL and MongoDB
The code changes in `mongo.py` and `postgres.py` update the comparison operators used in the PolishNotationToMongoDB and PostgresDatabaseManager classes. This includes adding support for new operators like `not_like`, `not_ilike`, `!like`, and `!ilike`, as well as updating the behavior of existing operators like `like` and `ilike`. These changes improve the flexibility and accuracy of query operations in both databases.

Ref # NVOMS-2707